refactor: 패키지 구조 및 경로 정리
