+++ src/main/java/com/mooddy/backend/feature/user/service/AuthService.java
@@ -6,9 +6,9 @@ import com.mooddy.backend.feature.user.dto.SignupRequest;
 import com.mooddy.backend.feature.user.dto.UserDetailResponseDto;
 import com.mooddy.backend.feature.user.domain.AuthProvider;
 import com.mooddy.backend.feature.user.domain.User;
-import com.mooddy.backend.global.exception.exception.AuthenticationException;
-import com.mooddy.backend.global.exception.exception.UserAlreadyExistsException;
-import com.mooddy.backend.repository.UserRepository;
+import com.mooddy.backend.global.exception.AuthenticationException;
+import com.mooddy.backend.global.exception.UserAlreadyExistsException;
+import com.mooddy.backend.feature.user.repository.UserRepository;
 import com.mooddy.backend.global.security.JwtService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
